refactor: 💡 sagas flow
