+++ src/modules/widgets/saga/chart/index.ts
@@ -1,3 +1,4 @@
 import { initializeChartWidget } from './initializeChartWidget';
+import { editChartWidget } from './editChartWidget';
 
-export { initializeChartWidget };
+export { initializeChartWidget, editChartWidget };
