+++ src/modules/widgets/utils/getEventCollections.ts
@@ -1,7 +1,6 @@
 import { Query } from '@keen.io/query';
 
-export const getEventCollections = (query: string | Query): string[] => {
-  if (typeof query === 'string') return [query];
+export const getEventCollections = (query: Query): string[] => {
   const { event_collection, analysis_type, steps } = query;
   if (analysis_type === 'funnel') {
     const collections: Set<string> = new Set();
